Implement curl_wrapper::download_files().
For this introduce file-handle and path in curl_handle_t
and add a new curl_handle_t::init(path) init-function.
So it is possible to tether a curl-handle to a file-handle
and both are freed automatically.
The file-handle stuff is optional for the use-case
downloading into a file.
It can be omitted for downloading into a buffer.
Refactor curl_wrapper::download_file() with it a bit.