Refactorization: mainly error-handling and remove png fake-header.
In main if there were less than 10% errors, try to redownload them
and if at the end there were only 1% overall errors, ignore them.
In the downloaded files check also for a PNG fake-header and remove it
(see pngfakeheader.h).
For this add file_util with read_file() and write-file() functionality.
If the m3u8-playlist only contains filenames (and no real relative paths),
then take the relativ path from the master-m3u8,
instead of only the protocol+server-part.
In curl_wrapper cancel downloading after 5 consecutive errors.
Add the url to curl_wrapper_error.
Extended verify_file() a bit by probing the failed download for html
and take the <title>-tag as the error-message if it is html.